printables-empire/requirements.txt
anthropic>=0.39.0
playwright>=1.40.0
playwright-stealth>=1.0.6
Pillow>=10.0.0
pyyaml>=6.0
pydantic>=2.0.0
